Add a Configure.pl probe for working #include <stdatomic.h>
C11 atomics are an optional C11 feature. mimalloc needs them, *and* it needs
a compiler conformant with a later correction to the spec, not the original
(buggy) C11 spec.

On some platforms we also might need to add -latomic to the linker flags to
pull in the necessary support library.