update-submodules.pl now handles linked worktrees created by git-worktree
`git worktree` allows multiple working trees attached to the same
repository, which is useful if one wants to have multiple development
projects checked out in parallel on the same machine. A side effect of its
implementation is that in a "linked working tree" `.git` is a *file*, not a
directory. `git` commands handle this just fine, but tools that assume that
they can test for a git checkout by simply looking for a directory `.git`
will fail.

Teach `tools/update-submodules.pl` that `.git` can be a file too.