Remove the fixed size allocator (FSA)
With the change to make using the mimalloc allocator the default, it
was discovered that it does such a good job that the FSA doesn't
provide any extra benefit. Therefore we're removing it to simplify
things in general. This will be a performance regression for the few
systems that can't use mimalloc, but those are very uncommon anyway.

Because the FSA had two jobs (fast allocations/frees for small amounts
of memory + an "at safepoint" mechanism to allow concurrent
modifications of things that aren't supposed to modified concurrently to
be detected and an error thrown instead of just segfaulting) and
mimalloc only replaces the first, the `*_at_safepoint()` functions of
the FSA had to be implemented in terms of the normal alloc/free
capabilities. That code was essentially copied from the FSA's
implementations.