Allocations for blocks of MVM registers needs to be 8 byte aligned
Various MVM_callstack_allocate_* functions allocate a single block of
memory, where the last section is used as an array of MVMRegister structs.
This section needs to be 8 byte aligned, as on some platforms the (real)
CPU's 64 bit loads from it need to be 8 byte aligned.

Most of these were correct already, but building on sparc32 reveals a few
that were not.

Also in `MVM_callstack_allocate_dispatch_record`, the allocation only needs
to be sizeof(MVMCallStackDispatchRun), as this is what is used, not
sizeof(MVMCallStackDispatchRecord). This hasn't been noticed previously,
because the latter struct is twice the size, so the bug simply resulting in
a bit of extra memory being allocated.