Use `atomic_load_explicit` on all compilers, not just clang
It turns out that on some architectures gcc generates less efficient code
without this, because for some CPUs it decides that the implicit memory
ordering that it should use is `memory_order_seq_cst`. Others default to
`memory_order_relaxed`, which is cheaper, and all that we need here.

Thanks to nine for asking what affect this code change would have, causing
me to investigate.