Fix segfault in nqp::hllboolfor
It currently doesn't check whether the fields it needs are present in
the HLLConfig returned from MVM_hll_get_config_for, which is fine because
all the names it's usually called with have existing and populated
HLLConfigs. But MVM_hll_get_config_for will create and return a mostly empty
HLLConfig if it *didn't* already exist for the given HLL name. So check and
throw an exception if the fields we need aren't populated.