Manually implement simplified strtoll for coerce_si
This way we don't have to allocate a new C string just to call strtoll
on it. Additionally, because we're always converting to base 10, we can
remove some of the checks that strtoll does.

There are two code paths, both use the same algorithm, but when the
MVMString is just storing ascii characters we can use its storage
directly, otherwise we have to go through a codepoint iterator.