Avoid reading beyond the end of the string in MVM_coerce_s_i
The recent refactoring to use `MVMCodepointIter` resulted in certain
strings (that are not valid numbers) causing the code to attempting to read
beyond the end of the string. For example, `nqp::coerce_si("-")` would
throw "«Iteration past end of grapheme iterator" instead of returning 0.

From inspection I believe that the code for MVM_STRING_GRAPHEME_ASCII
could also read the byte one beyond the end of the string for certain cases,
but have not been able to test this.

Also, the code was terminating loops where ord == 0 (ie ASCII NUL) which is
no longer needed, now that we terminate the loop based on comparing the count
of read ASCII with `strgraphs`, or whether `MVM_string_ci_has_more` returns
true.

Hence refactor both branches to the same structure, with explicit end tests
and early returns for the various corner cases (seemingly valid string
terminates before any digits are found).