Fix segfault after capture-replace-literal-arg with different kind
A comment above MVM_capture_replace_arg already said: "The callsite argument
type is expected to be the same". Sadly comments are easy to miss and when
arguments are replaced with values of different kinds, we can end up with e.g.
a literal number for a callsite argument with kind obj and may try to
dereference what we wrongly believe to be a pointer.

Fix by actually enforcing this rule. Fixes the segfault reported in GH #1656