Deserialize bigints without allocating a temporary copy of the string
Use MVMCodepointIter directly to read characters from the string
representation, instead of creating a temporary copy of the ASCII bytes
(which uses MVMCodepointIter internally for the common cases).

Also, as we're only handing strings with radix 32, use `mp_mul_2d` instead
of `mp_mul_d`, as the former is slightly more efficient.