Support all kinds of rw args on native calls in spesh
So far, we've only supported the larger integer types for rw args when speshing
native calls. Extend this support to all sizes, unsigned integers and pointers.

Since CPointer objects get passed as-is, not as native refs, we need to update
them in-place via sp_bind_i64 instead of assigning into the ref.