Don't use MoarVM to free memory it didn't allocate
We're not currently using MoarVM allocator functions in GMP, so don't use
MoarVM's free functions either. This didn't use to be a problem, but now
that we optionally use mimalloc, it can be. The optimal fix is to tell GMP
about our allocation functions, but this will work for now.