Fix frame walker finding wrong value after multi level inlining
The frame walker expects a spesh cand's inlinees to be sorted depth first. This
has the benefit of the walker being able to just iterate over the inlinees and
take the first one that covers the current bytecode position.

In the failing case we'd inlined Map.keys which is just:
Seq.new(Rakudo::Iterator.Mappy-keys(self)). Originally the spesh cand for keys
did not have any inlines, so none could get stacked up front. After inlining we
ran optimize_bb over the inlined graph to get a more optimized version.
optimize_bb would then encounter those dispatches, try to optimize them and
succeed at inlining. But those inlinees got added to the table after the
original inlinee for keys. This caused the frame walker to think that we're
a frame out of the actual one, which made it return the wrong value for
::?CLASS

This is acutally a new-disp regression. Commit bc98cb9467ab2d5d408dff841f39ae87
originally introduced optimization of inlined graphs after inlining but
intentionally did not optimize calls in the inlined graph to avoid the problem
described above. When the code got translated for new-disp, this safety measure
was omitted. The fix is to simply re-instate the inline-blocker (but without
blocking the optimization to pick a spesh cand).

A possible future optimization could be to allow for inlining but be smarter
about allocating the inline indexes.

Fixes Rakudo issue #4655