Fix spesh of named uint parameters
Handle the newly added `*_u` versions of named parameters ops. Now
something like `sub a(uint :$a!) { $a ?? $a + 1 !! 0 }; my uint $a = 1;
for ^1_000_000 { $a = a(:$a); }; say $a;` no longer dies with `Unexpected
named argument 'a' passed`, even though it finishes fine with a smaller
number of iterations in the loop.