Fix speshing of param_on_o
This case was added in https://github.com/MoarVM/MoarVM/pull/1681, but
isn't necessary and causes spesh to not turn it into an sp_* op, and then
since it can't be jitted either it was causing a bail. So remove the
body of the case, but leave it in so nobody else thinks it was
accidentally left out.