Fix errata in docs/debug-server-protocol.md
The debug server protocol document did not match actual server behavior in
several instances, and had various typos and missing details as well.

Because all of these errors were in the document, *not* in the code, fix
these errors with a pure errata update rather than a protocol minor version
change.

See the top of the diff for a list of all fixed errata.