Run finalizers even with tc->last_handler_result
Finalizers can run at pretty much any return point, and it is possible
that they could use exception handling that thus end up replacing the
content of `tc->last_handler_result`, affecting the code running on the
thread before it was asked to run finalizers. The initial solution for
this was to not run finalizers if `tc->last_handler_result` was set, on
the assumption that this would only be an occasional and transient
condition.

In reality, programs exist that end up with it set pretty much all the
time, with the upshot that finalizers would never be executed. For
example, https://github.com/libxml-raku/LibXML-raku/issues/76 ended up
leaking as a result of DESTROY never being invoked.

Instead of refusing to run finalizers with `tc->last_handler_result`
set, save the value before running finalizers and then restore it
afterwards. This fixes the leak observed in the LibXML module.