Fallback resolvers for compunits and syscall for setting them
When we run code at BEGIN time the actual lexical environment the code
is supposed to run in hasn't been set up yet. Previously we got around
that by compiling a wrapper frame, containing all lexicals visible from
the code we run. This frame is rather expensive and we want to replace
it with an on-demand resolver that's part of the compiler (i.e. HLL
code)