Reimplement the Semaphore REPR by atomic ops
libuv is a handicap for us in this REPR because it lacks basic
capabilities of libpthread semaphores. In particular, we need
`sem_getvalue`, whose is preventing us from cloning or serializing
semaphores as is possible with `ReentrantMutex`, which combined with
the nqp atomic ops' dependence on containers, prevents us from having a
simplistic thread-safe counter in Rakudo's MOP. We can avoid any
unnecessary synchronization when C11 atomic ops are supported, which
gives us acquire/release memory ordering as opposed to full barrier
spam, so we wind up with operations that align with how an atomic
counter would typically be used. While we're here, a semaphore should
never box a negative int.