Reset errno to error of latest call to mkdir
The old code would overwrite the error from mkdir with that from the
following call from uv_fs_stat. Since "errno" is used for error
reporting in MVM_dir_mkdir this could lead to strange error messages.

For instance, the attempt to create a directory when there is already
a file at the given path, currently leads to this output:

  $ ./rakudo-m -e 'mkdir("existing-file")'
  Failed to create directory '/path/to/existing-file' with mode '0o777': Failed to mkdir: Success
  in block <unit> at -e line 1

The new version mentions the underlying problem:

  $ ./rakudo-m -e 'mkdir("existing-file")'
  Failed to create directory '/path/to/existing-file' with mode '0o777': Failed to mkdir: File exists
  in block <unit> at -e line 1