Speedup VMArray's deserialize()...
by moving a switch outside of a loop. This means we have to duplicate
the loop inside each case statement, but the compiler generates better
code. For `raku -e ''`, we deserialize almost 90k VMArrays. With
`MVM_SPESH_DISABLE=1` we go from ~626k to ~623k instructions and with
`MVM_SPESH_BLOCKING=1` from ~1.135b to ~1.132b instructions.