Work around "const_iX NYI" after getlexstatic_o with fallback resolver
getlexstatic_o and sp_getlexstatic_o did not take into account that
MVM_frame_find_lexical_by_name may now dispatch if a fallback resolver
is in use.

Note: this fix is _not_ complete as it completely ignores that we have
that dispatcher mechanism and that getlexstatic_o is already implemented
using it. Thus we store the ostensibly temporary VMNull in the dispatch
cache. A repeated call will therefore get the wrong result. Alas the
fallback resolver is only in use at BEGIN time and repeated execution
there is rare and I currently lack the smarts for a proper fix.