Convert all dir functions to just use libuv...
instead of doing our own `#define`s for Windows support.

Almost all the logic is the same. However, libuv removed '.' and '..'
from readdir results, and since Rakudo (and roast) expects them, this
will require a minor change in Rakudo. The JVM also doesn't return those
for its readdir implementation, so there's already code that deals with
that in Rakudo in `#?if jmv`s, which will just need to be adjusted to
be `#?if !js`.