Enable c11 atomics on msvc 17.8+
Starting with MSVC 17.8 Preview 2, c11 atomics can be used to build
with a msvc toolchain. To get c11 atomics to work we need to do a
few things:

* Pass the /experimental:c11atomics flag until msvc implements
  locking atomics
* Use c17 instead of c++latest, as the microsoft C11 atomics ops
  implementation doesn't work if it is in _cplusplus mode
* Fix a couple spots where the compiler complained, presumably
  because of the change from c++latest to c17

I don't know how well this works as an unrelated failure prevents
me from testing it. But it does compile now  with --c11-atomics
when using the latest preview build of MSVC.

See:
https://devblogs.microsoft.com/cppblog/c11-threads-in-visual-studio-2022-version-17-8-preview-2/