Support long paths on Windows
On Windows if you want path parts longer than MAX_PATH then you
need to use DOS paths, i.e. absolute paths prefixed with \\?\.
Additionally, using DOS paths does not work with / so those slashes
need to be normalized when prefixing paths with \\?\. This adds
the aforementioned prefix and normalization on Windows, allowing
long paths to be used on that OS.

Note this is a Proof of Concept and I haven't spent any time thinking
about the design or what I may have otherwise broke.