Restore the use of S_ISDIR in mkdir logic
In a previous commit the windows specific code for mkdir_p was
removed such that all platforms would use the same logic. However,
the non-Windows logic had a S_ISDIR check that I temporarily
removed due to it not being available on Windows.

This adds the S_ISDIR macro on Windows, and readds the original
S_ISDIR check to our mkdir logic.