Disable invalid parameter warnings
`/experimental:c11atomics` is needed for cutting edge msvc toolchains to use c11 atomics. When using this parameter on older msvc toolchains it would emit a warning about an unknown parameter. This disables the warning for unknown parameters until some time in the future when the `/experimental` parameter is no longer needed.