Only prefix paths longer than MAX_PATH
Absolute paths of any length can use the \\?\ prefix, so we
initially implemented things without checking the length. However,
the native nqp and rakudo runners currently use absolute paths that
contain `..` as a path part which is literal (i.e. not the parent
directory) when used with the \\?\ prefix. This checks the path
length to see if it is longer than MAX_PATH so we avoid the issue
with the nqp/rakudo runners in most situations while still getting
some immediate relief to other problems stemming from Windows path
length issues.