Use atomic operations for on SerializationReader's working flag
This flag is read from different threads than it is written to. This
means that when using a plain integer, a thread could read an outdated
value from a CPU cache leading to it believing that no deserialization
is running when it actually is. This in turn could cause us to read from
half-deserialized objects.

lizmat++ for all the heavy debugging work on this