(twoway_memmem_uint32) Fix memcmp call
This resolves Rakudo Issue #5550 (Simple literal-only regex fails).

Apparently the previous behavior was Good Enough for 6 years of
regexing, but it contained a significant bug: `memcmp` was being
given 32-bit byte strings but was instructed only to compare the
first `ms+1` bytes.

Surprisingly, it seems that the overall algorithm still does the
right thing for the vast majority of cases. However, when the
haystack / needle combination is of a certain patternality, we
get an obviously wrong failure.

Haystack = Xabddcabaacab
Needle   = abaacab

Note that the characters don't matter, just the ordering and
combination.