Don't lose return value with LEAVE phasers
When using &return an exception is thrown and the return value saved in
`tc->last_payload`. LEAVE phasers interupt the stack unwind and call the
LEAVE code block. Code in there can and often does make use of
`tc->last_payload` via &return or other means. This overwrites the original
value which is then lost. Fix by saving / restoring the original value in
the special return stack frame. That frame is added to the call stack
before calling the LEAVE code block and the value can be restored once we
are back to that stack frame.