Implement return prioritization
Calling &return triggers a stack unwind which can cause code (e.g. in
`LEAVE` phasers) to run. That code can also call `&return`. The returns
conflict, if the later return unwinds over the return continuation
marker of the former return in the call stack, i.e. the `LEAVE` frame.
In such situations one wants the return to win that unwinds more of the
stack. This is the behavior implemented by this change. Previously the
later return always won.

This implements the behavior discussed in Raku/problem-solving#417.