optimize elems in rare cases
such as when we have the create or fastcreate for the elems directly
in front of it and nothing changes it. This happens when we optimize
a "get slurpy named arguments" and we know what arguments there
are.

Currently we only optimize elems to a constant 0 if nothing is put
into the object, but in the future we could get the exact number.