Add MVM_string_decodestream_get_up_to_chars
Allows us to ask for an amount of characters, and get something out
of the decodestream even when there's not as much to get as we ask
for.

IO::Handle's readchars method by default tries to
get $*DEFAULT-READ-ELEMS from the decoder and won't return
anything until that amount of chars has become available from
repeated reads.

This change brings MVM_decoder_take_chars behaviour in line with
the comment in front of it ("or all if there is not enough") and fixes
both the readchars and Supply methods on $*IN with an encoding
blocking until a lot of input has been read.

IO::Handle.lines.Supply already did what you wanted, and if you unset
the encoding on $*IN first, .Supply.lines also works. Now both work.