Split string's "any" union field into "any" and "any_ptr"
In some spots we were using "any" to be able to copy over the
contents of a string no matter what storage type it was, in others
we wanted to refer to any of the pointers in it without having to
specify which (even though it doesn't actually matter).

However, on 32bit systems, the old any field (void *) was only half
as long as the in_situ_8 and in_situ_32 fields, which causes us to
miss up to half of the graphemes, for example when reaching
a string in a grapheme iterator.