Fix NAN/INFINITY definitions to work on Windows 11 SDK
As of a recent Windows 11 SDK version NAN/INFINITY are no longer
compile time constants. That update caused the moarvm build to
fail with a "error C2099: initializer is not a constant" error.
This updates our NAN and INFINITY definitions when using MSVC to
avoid assigning the (now) non compile time constants to a static
const.

Fixes #1837