when a hash oopses, heap snapshot and look for refs
The idea is that if a hash was concurrently changed without
proper locking, then we may still find a reference to the
hash that exploded in other thread's memory, in particular
the registers and lexicals of its frames.

However, outputting this information completely unfiltered
is not helpful: it is a huge amount, and any thread that
has any frame on its stack that has an outer frame that
has a lexical pointing at our object will show up here,
which is a very low strength signal for whether the thread
is manipulating the hash, or even looking at it.

The graphviz output from this code is not much better
than the terminal output, thanks to long labels on the
edges as well as far too many incident edges on some
of the nodes, especially close to our object.