Counteract drift in timers with repeat
With regular repeating timers in libuv, we accumulate
a small but significant amount of extra time with
every tick.

We only get to supply millisecond precision values
for the timeout and repeat values, so we can
only correct the ticks by one millisecond or more,
and we only notice when the difference is above that
same amount as well.