debugserver: prevent GC oops when stepping past a gc run
in order to refer to a frame as a collectable, it has to live in
the heap, it must not be on the callstack.

Instead of making sure that we clean the step_mode_frame whenever
something happens to make it become invalid or anything complicated
like that, for now we just force the stack frame onto the heap,
which is not optimal from a performance standpoint, but safe.