More helpful error when a VMException object is serialized
Until now, we just get the generic "missing serialize function" error,
but to an end user, it is not meaningful or actionable.

Instead, we give VMException a serialization function that outputs an
error message that explains at least a little bit what went wrong, and
more importantly spits out the exception's message as well as its
backtrace.