Fix OS-is-windows check that doesn't work in this file
Switch "linux" vs "non-linux" to "windows" vs "non-windows"

The code for mac should be the same as linux, not the same
as windows. We do not use powershell nor Visual Studio on mac.

don't error out if catchsegv doesn't exist (like on mac)

Use macos-14 by default and -13 for the "old OS" job