Maybe optimize MVM_string_latin1_decode()
Reduce allocations in MVM_string_latin1_decode

Check up-front if the result is going to be 8 or 32 bit and have
different paths for each case. Also, check if the result could be
in-situ before starting decoding, so we can write directly into the
in-situ buffer in that case instead of alloc+copy+free.

Both gcc and clang produce code that's faster than actually breaking out
of the loop early. At least on x86-64 it's being vectorized, I can't
tell exactly what's happening on aarch64.