Make accesses to spesh arg guard atomic (#1903)
* Make accesses to spesh arg guard atomic

On ARM and other platforms with less strict memory ordering guarantees
than x86, the step that spesh does to switch the pointer around to the
newly created spesh arg guard can be seen by another thread before the
contents of the arg guard are seen, causing the thread to jump into
uninitialised "code". Until recently, this just caused freezing or
random behaviour. Then we got a panic or oops in moar when the code
"jumped" out of bounds, which is both better and worse.

This is a real fix for the issue.

Need to check if the performance impact on x86_64 is significant.
In that case, for that arch, we could conditionally use regular access
instead of atomic ops. Alternatively the switch-over of the spesh arg
guard pointer could be delayed, but then I don't know how to guarantee
the new spesh arg guard contents are really visible at that time.