Add type specialized VMArray atpos implementations
For the numeric type ops (`*_[iun]`) I only added the most common sizes
(64-bit) and check beforehand that the array's slot type matches,
otherwise just go through the specialized-for-core-types function. For
the string and object type ops we can just go directly to the single new
version.

Then call these new versions directly when we can elsewhere in the code.