Avoid segfault in GC with nqp::objectid
When we generate an objectid, we just reserve some space in gen2 and use
the pointer as the stable object id. However, the object does not
immediately get moved there. Thus that space remains uninitialized
(zeroed out). When the GC collects that space (because the object went
away) and tries to run finalizers, we segfault trying to get that
object's REPR. Avoid by checking whether we have an STABLE poitner in
the first place.

To be honest, this doesn't really look like an actual fix and feels more
like a workaround. We probably should move objects to gen2 instead when
giving them an id. But at least it avoids those crashes for now.