Less memory waste binslurp and normalizecodes (#1938)
* In read_fhb aka nqp::readfh don't keep too-large buffers around.

We were allocating as much space as the caller was asking for,
but in rakudo for slurp we are passing 1 megabytes as the size by
default.

* normalizecodes: allow re-using codepoints buffer without leaking

Passing a buffer object in that already had contents used to drop the
buffer on the floor without freeing it.

Also, we weren't setting ssize, which could in theory make vmarray misbehave.