Simplify alignof definition
The old one gave all sorts of warnings for me (on clang, with the
fallback definition). C11 provides a built-in solution in stdalign.h,
which I'm using now instead.

Long term getting rid of ALIGNOF and just use alignof instead is
probably better but I've left that for the future.