Update ucd2c.pl for Unicode 17
This involves two fixes: the first is that DerivedCoreProperties.txt now
includes enumerated properties alongside binary properties, and thus
handling for that had to be added. At present it just has hardcoded
support for the new Indic_Conjunct_Break property, but it shouldn't be
too hard to generalize if ever needed.

The second fix is simply to include support for a new CJK Ideograph
Extension.