Add some easy enumeration properties to ucd2c.pl
One of these, Indic_Syllabic_Category, is required for proper grapheme
handling, so obviously it needs to be added in. The other enumerated
property files added were simply easy enough to throw in at the same
time.

There's quite a bit more work to do on adding in all the property data
Unicode provides, but for now the primary focus is on updating to
Unicode 17. Afterwards we can focus on improving support for that new
version of Unicode.