Fix grapheme boundary searching for Unicode 17.
This introduces a long-needed change to how MoarVM searches for grapheme
clusters, and it now properly handles all graphemes as per the default
rules outlined in UAX#29, revision 47. This required a completely new
function, because the old one was designed for a much different set of
rules, and has led to appropriate changes elsewhere.

The new grapheme boundary function finds the end of a grapheme cluster
from a given starting position, which allows for a much more natural
implementation of the rules, and should be easier to augment if and when
new kinds of graphemes are added to the default rules in the future. The
old approach, where each invocation of the search function could only
look one codepoint ahead and behind, would've required a less
approachable state machine to work with the more modern rules.

When it comes to the case of joining two NFG strings and checking to see
if they need renormalization around the join, we fortunately only have
to examine the graphemes immediately adjacent to the join point, as
before, though it did have to be changed to work with the new search
function. The join-checking function can be used for any two graphemes,
but for now a minimal change was made, and it's only used to check
non-synthetic codepoints (as was the case beforehand).