Stop punishing concatenations around CRLF.
Thanks to japhb++ for spotting this in review. Turns out we've been
accidentally making concatenations where one side of the join is a CRLF
grapheme get renormalized, possibly for a long time, even though the
intent was the opposite. A simple enough fix, but should speed up any
nqp or Rakudo code that happens to hit this code path a bunch.