Fix minor, unrealized bug in MVM_unicode_normalize_nfg_breaks
This function was written to potentially be used with synthetic
codepoints in the future, but as written would've incorrectly unwrapped
UTF-C8 codepoints (when they should be left alone, and always stay
broken off from other graphemes).

Currently, the one use of this function only uses it on non-synthetic
codepoints, so this commit doesn't change any currently reachable
behavior. The fix is really just because I spotted it and couldn't bear
to leave it to bite someone in the future.